Refactor: Moving entities to specified packages. Added eduVirt DS base config
