+++ src/main/java/pl/lodz/p/it/eduvirt/entity/eduvirt/Course.java
@@ -1,30 +1,29 @@
-package pl.lodz.p.it.eduvirt.entity;
+package pl.lodz.p.it.eduvirt.entity.eduvirt;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import lombok.*;
 import org.hibernate.proxy.HibernateProxy;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-@Getter
 @Setter
+@Getter
+@Table(name = "i72_course")
 @Entity
-@Table(name = "i72_resource_group_pool")
-public class ResourceGroupPool extends HistoricalData {
+public class Course extends AbstractEntity {
     @Column(name = "name", nullable = false, length = 100)
     private String name;
-    @Column(name = "max_rent", nullable = false)
-    private int maxRent;
-    @Column(name = "grace_period", nullable = false)
-    private int gracePeriod;
-
-    @OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    private List<ResourceGroup> resourceGroups = new ArrayList<>();
+    @Column(name = "description", nullable = false, length = 1000)
+    private String description;
+    @OneToMany
+    private List<ResourceGroupPool> resourceGroupPools;
 
     @Override
     public final boolean equals(Object o) {
@@ -33,8 +32,8 @@ public class ResourceGroupPool extends HistoricalData {
         Class<?> oEffectiveClass = o instanceof HibernateProxy hibernateProxy ? hibernateProxy.getHibernateLazyInitializer().getPersistentClass() : o.getClass();
         Class<?> thisEffectiveClass = this instanceof HibernateProxy hibernateProxy ? hibernateProxy.getHibernateLazyInitializer().getPersistentClass() : this.getClass();
         if (thisEffectiveClass != oEffectiveClass) return false;
-        ResourceGroupPool that = (ResourceGroupPool) o;
-        return getId() != null && Objects.equals(getId(), that.getId());
+        Course course = (Course) o;
+        return getId() != null && Objects.equals(getId(), course.getId());
     }
 
     @Override
