+++ src/main/java/pl/lodz/p/it/eduvirt/entity/eduvirt/ResourceGroupPool.java
@@ -1,29 +1,30 @@
-package pl.lodz.p.it.eduvirt.entity;
+package pl.lodz.p.it.eduvirt.entity.eduvirt;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.proxy.HibernateProxy;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-@Setter
 @Getter
-@Table(name = "i72_course")
+@Setter
 @Entity
-public class Course extends AbstractEntity {
+@Table(name = "i72_resource_group_pool")
+public class ResourceGroupPool extends HistoricalData {
     @Column(name = "name", nullable = false, length = 100)
     private String name;
-    @Column(name = "description", nullable = false, length = 1000)
-    private String description;
-    @OneToMany
-    private List<ResourceGroupPool> resourceGroupPools;
+    @Column(name = "max_rent", nullable = false)
+    private int maxRent;
+    @Column(name = "grace_period", nullable = false)
+    private int gracePeriod;
+
+    @OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    private List<ResourceGroup> resourceGroups = new ArrayList<>();
 
     @Override
     public final boolean equals(Object o) {
@@ -32,8 +33,8 @@ public class Course extends AbstractEntity {
         Class<?> oEffectiveClass = o instanceof HibernateProxy hibernateProxy ? hibernateProxy.getHibernateLazyInitializer().getPersistentClass() : o.getClass();
         Class<?> thisEffectiveClass = this instanceof HibernateProxy hibernateProxy ? hibernateProxy.getHibernateLazyInitializer().getPersistentClass() : this.getClass();
         if (thisEffectiveClass != oEffectiveClass) return false;
-        Course course = (Course) o;
-        return getId() != null && Objects.equals(getId(), course.getId());
+        ResourceGroupPool that = (ResourceGroupPool) o;
+        return getId() != null && Objects.equals(getId(), that.getId());
     }
 
     @Override
