+++ src/main/java/pl/lodz/p/it/eduvirt/entity/eduvirt/reservation/keys/ClusterMetricKey.java
@@ -1,9 +1,10 @@
-package pl.lodz.p.it.eduvirt.entity.reservation.keys;
+package pl.lodz.p.it.eduvirt.entity.eduvirt.reservation.keys;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import pl.lodz.p.it.eduvirt.entity.eduvirt.general.Metric;
 
 import java.util.UUID;
 
@@ -11,8 +12,8 @@ import java.util.UUID;
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode
-public class ClusterNetworkKey {
+public class ClusterMetricKey {
 
     private UUID clusterId;
-    private UUID networkId;
+    private Metric metric;
 }
