+++ src/main/java/pl/lodz/p/it/eduvirt/entity/eduvirt/reservation/keys/ReservationLogKey.java
@@ -1,10 +1,10 @@
-package pl.lodz.p.it.eduvirt.entity.reservation.keys;
+package pl.lodz.p.it.eduvirt.entity.eduvirt.reservation.keys;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import pl.lodz.p.it.eduvirt.entity.reservation.Reservation;
+import pl.lodz.p.it.eduvirt.entity.eduvirt.reservation.Reservation;
 
 import java.util.UUID;
