+++ src/main/java/pl/lodz/p/it/eduvirt/repository/ovirt/VnicProfileRepository.java
@@ -1,8 +1,8 @@
-package pl.lodz.p.it.eduvirt.repository;
+package pl.lodz.p.it.eduvirt.repository.ovirt;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pl.lodz.p.it.eduvirt.entity.network.VnicProfilePool;
+import pl.lodz.p.it.eduvirt.entity.ovirt.VnicProfilePool;
 
 import java.util.UUID;
